feat(explore): Making group by rows draggable with numeric tags
