+++ static/app/views/explore/toolbar/toolbarGroupBy.tsx
@@ -1,89 +1,44 @@
-import {useCallback, useMemo} from 'react';
+import {useCallback, useEffect} from 'react';
+import styled from '@emotion/styled';
 
-import {Button} from 'sentry/components/button';
-import type {SelectOption} from 'sentry/components/compactSelect';
-import {CompactSelect} from 'sentry/components/compactSelect';
 import {IconAdd} from 'sentry/icons/iconAdd';
-import {IconDelete} from 'sentry/icons/iconDelete';
 import {t} from 'sentry/locale';
+import {space} from 'sentry/styles/space';
+import {defined} from 'sentry/utils';
 import {useGroupBys} from 'sentry/views/explore/hooks/useGroupBys';
-import type {Field} from 'sentry/views/explore/hooks/useSampleFields';
 
+import {ColumnEditor} from '../components/columnEditor';
 import {useSpanTags} from '../contexts/spanTagsContext';
+import {useDragNDropColumns} from '../hooks/useDragNDropColumns';
+import {useResultMode} from '../hooks/useResultsMode';
 
-import {
-  ToolbarHeader,
-  ToolbarHeaderButton,
-  ToolbarLabel,
-  ToolbarRow,
-  ToolbarSection,
-} from './styles';
+import {ToolbarHeader, ToolbarHeaderButton, ToolbarLabel, ToolbarSection} from './styles';
 
 interface ToolbarGroupByProps {
   disabled?: boolean;
 }
 
 export function ToolbarGroupBy({disabled}: ToolbarGroupByProps) {
-  const tags = useSpanTags();
+  const numberTags = useSpanTags('number');
+  const stringTags = useSpanTags('string');
+  const [resultMode] = useResultMode();
 
   const {groupBys, setGroupBys} = useGroupBys();
 
-  const options: SelectOption<Field>[] = useMemo(() => {
-    // These options aren't known to exist on this project but it was inserted into
-    // the group bys somehow so it should be a valid options in the group bys.
-    //
-    // One place this may come from is when switching projects/environment/date range,
-    // a tag may disappear based on the selection.
-    const unknownOptions = groupBys
-      .filter(groupBy => groupBy && !tags.hasOwnProperty(groupBy))
-      .map(groupBy => {
-        return {
-          label: groupBy,
-          value: groupBy,
-          textValue: groupBy,
-        };
-      });
+  const {editableColumns, updateColumnAtIndex, deleteColumnAtIndex, swapColumnsAtIndex} =
+    useDragNDropColumns({columns: groupBys});
 
-    const knownOptions = Object.keys(tags).map(tagKey => {
-      return {
-        label: tagKey,
-        value: tagKey,
-        textValue: tagKey,
-      };
-    });
-
-    return [
-      // hard code in an empty option
-      {label: t('None'), value: '', textValue: t('none')},
-      ...unknownOptions,
-      ...knownOptions,
-    ];
-  }, [groupBys, tags]);
+  useEffect(() => {
+    setGroupBys(editableColumns.map(({column}) => column).filter(defined));
+  }, [editableColumns, setGroupBys]);
 
   const addGroupBy = useCallback(() => {
     setGroupBys([...groupBys, '']);
   }, [setGroupBys, groupBys]);
 
-  const setGroupBy = useCallback(
-    (i: number, {value}: SelectOption<Field>) => {
-      const newGroupBys = groupBys.slice();
-      newGroupBys[i] = value;
-      setGroupBys(newGroupBys);
-    },
-    [setGroupBys, groupBys]
-  );
-
-  const deleteGroupBy = useCallback(
-    (index: number) => {
-      const newGroupBys = groupBys.filter((_, orgIndex) => index !== orgIndex);
-      setGroupBys?.(newGroupBys);
-    },
-    [setGroupBys, groupBys]
-  );
-
   return (
     <ToolbarSection data-test-id="section-group-by">
-      <ToolbarHeader>
+      <StyledToolbarHeader>
         <ToolbarLabel disabled={disabled}>{t('Group By')}</ToolbarLabel>
         <ToolbarHeaderButton
           disabled={disabled}
@@ -93,28 +48,22 @@ export function ToolbarGroupBy({disabled}: ToolbarGroupByProps) {
           aria-label={t('Add Group')}
           icon={<IconAdd />}
         />
-      </ToolbarHeader>
+      </StyledToolbarHeader>
       <div>
-        {groupBys.map((groupBy, index) => (
-          <ToolbarRow key={index}>
-            <CompactSelect
-              searchable
-              disabled={disabled}
-              options={options}
-              value={groupBy}
-              onChange={newGroupBy => setGroupBy(index, newGroupBy)}
-            />
-            <Button
-              borderless
-              icon={<IconDelete />}
-              size="zero"
-              disabled={disabled || (groupBys.length <= 1 && groupBy === '')}
-              onClick={() => deleteGroupBy(index)}
-              aria-label={t('Remove Group')}
-            />
-          </ToolbarRow>
-        ))}
+        <ColumnEditor
+          disabled={resultMode === 'samples'}
+          columns={editableColumns}
+          onColumnChange={updateColumnAtIndex}
+          onColumnDelete={deleteColumnAtIndex}
+          onColumnSwap={swapColumnsAtIndex}
+          stringTags={stringTags}
+          numberTags={numberTags}
+        />
       </div>
     </ToolbarSection>
   );
 }
+
+const StyledToolbarHeader = styled(ToolbarHeader)`
+  margin-bottom: ${space(1)};
+`;
